+++ src/engine/engineVariables.ts
@@ -30,6 +30,7 @@ export class EngineVariables {
         e2eTestDistFolder: string;
         reportsFolder: string;
         packageFolder: string;
+        buildFolder: string;
 
         assetsFolder: string;
         assetsSourceFolder: string;