+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.8.0
 * Added platform packager web
+* Added platform packager electron with icon support for win32/darwin/linux
+* Added version gulp tasks
 * unite.json directories configuration changes to have www sub property
 * node_modules path rationalised in config
 * Fixed safari_pinned_tab color in meta headers